requirements.txt
flask
pypugjs
jinja2
libsass
sqlalchemy
psycopg2
python-dotenv
flask-sqlalchemy